Update atomes.spec
Hi Sébastien,
I took a quick look and made some corrections to the spec file cosmetic (frenchisms, weird ligatures, etc.) and others.
Your changelog entries need to have the  (E)VR information at the end. When you make a new build, but there is no code change in the package upstream, you just increment the release, both in the "Release:" field and in the changelog.
Also, note that you are writing the changelog from the perspective of the packager and not the upstream developer (which is often confusing when you are both). This means that you note things like new program versions, a patch you may have included, a rebuild against a newer version of a dependency etc..
While I was able to replace `make install` with the `%make_install` macro, I could not replace `make` with `%make_build` in the build phase. On my computer, the macro evaluates to
$ rpm --eval %make_build
/usr/bin/make -O -j8 V=1 VERBOSE=1
I'm guessing that something might be wrong with parallel builds, as I would always get a complaint about a missing file - I think parameters.mod or something like that.

These changes should make the linter complain a bit less, if you have any questions, don't hesitate to ask.